+++ src/smartquotebot/bot.py
@@ -1,9 +1,11 @@
 import asyncio
-import discord
-from discord.ext import commands
-from startup import app_ctx
 
-intents = discord.Intents.default()
+from discord import Intents
+from discord.ext import commands
+
+from smartquotebot.startup import app_ctx
+
+intents = Intents.default()
 intents.message_content = True
 
 bot = commands.Bot(command_prefix=app_ctx.config.bot.command_prefix, intents=intents)
@@ -11,7 +13,7 @@ bot = commands.Bot(command_prefix=app_ctx.config.bot.command_prefix, intents=int
 
 async def load_extensions():
     try:
-        await bot.load_extension("cogs.quotes")
+        await bot.load_extension("smartquotebot.cogs.quotes")
         app_ctx.logger.info("Loaded extension: cogs.quotes")
     except Exception:
         app_ctx.logger.exception("Failed to load extension: cogs.quotes")
